# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "URL Shortener API",
    "version": "1.0.0",
    "description": "API para acortar URLs. Proyecto demostrativo de un flujo CI/CD completo."
  },
  "paths": {
    "/api/urls": {
      "post": {
        "summary": "Crear una URL corta",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": { "type": "string", "example": "https://example.com" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "URL creada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "shortCode": { "type": "string", "example": "abc1234" },
                    "shortUrl": { "type": "string", "example": "http://localhost:3000/abc1234" }
                  }
                }
              }
            }
          },
          "400": { "description": "URL inválida" }
        }
      }
    },
    "/api/urls/{shortCode}": {
      "get": {
        "summary": "Obtener metadatos de una URL corta",
        "parameters": [
          {
            "name": "shortCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Metadatos de la URL" },
          "404": { "description": "No encontrado" }
        }
      }
    },
    "/{shortCode}": {
      "get": {
        "summary": "Redirigir a la URL original",
        "parameters": [
          {
            "name": "shortCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "301": { "description": "Redirección a la URL original" },
          "404": { "description": "No encontrado" }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": { "description": "Servicio disponible" }
        }
      }
    }
  }
}